Sending a GET Request
The code above is a function called App that sets an initial empty movies state using the useState. The fetchMoviesHandler function is used to fetch movies from the https://swapi.dev/api/films/ API. After getting the response data, it transforms the data into a more suitable format and updates the state of movies using the setMovies function. The map function is used to create a new movie object with id, title, openingText and releaseDate properties. Finally, the movies state is updated using setMovies with the new array of transformed movies.